requirements.txt
pyserial==3.4
python-can==3.3.4